+++ Services/Interfaces/IContactService.cs
@@ -5,9 +5,9 @@ namespace CashOut.Services.Interfaces
     public interface IContactService
     {
         public List<Contact> GetAll();
-        public Contact GetById(int id);
+        public Contact GetById(long id);
         public object Add(Contact contact);
-        public int Delete(int id);
+        public int Delete(long id);
         public int Update(Contact contact);
         public Contact GetByPhone(string phone);
     }
